requirements_dev.txt
invoke==1.6.0
twine==3.4.2
zest.releaser==6.22.1
bumpversion==0.6.0
django-model-utils==4.1.1
wheel==0.37.0
